requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
dnspython==1.16.0
greenlet==0.4.15
idna==2.8
monotonic==1.5
multidict==4.5.2
python-engineio==3.9.0
python-socketio==4.3.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
websocket-client==0.56.0
yarl==1.3.0
autopep8==1.4.4
Babel==2.7.0
Brotli==1.0.7
certifi==2019.6.16
chardet==3.0.4
cpplint==1.3.0
cycler==0.10.0
deap==1.3.0
Delorean==1.0.0
docopt==0.6.2
futu-api==3.10.2
helpdev==0.6.10
humanize==0.5.1
ibapi==9.75.1
idna==2.8
importlib-metadata==0.18
kiwisolver==1.1.0
matplotlib==3.1.0
mongoengine==0.18.2
msgpack==0.6.1
numpy==1.16.4
pandas==0.24.2
pbr==5.2.0
peewee==3.9.6
pipenv==2018.11.26
protobuf==3.8.0
psutil==5.6.3
psycopg2==2.8.3
pyasn1==0.4.4
pycodestyle==2.5.0
pycryptodome==3.8.2
pymongo==3.8.0
PyMySQL==0.9.3
pyparsing==2.4.0
PyQt5==5.12.3
PyQt5-sip==4.19.17
pyqtgraph==0.10.0
PySocks==1.7.0
python-dateutil==2.8.0
python-rapidjson==0.7.2
pytz==2019.1
QDarkStyle==2.7
requests==2.22.0
rqdatac==2.1.0
rsa==4.0
scipy==1.3.0
seaborn==0.9.0
simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.3.5
stevedore==1.30.1
stomp.py==4.1.22
TA-Lib==0.4.17
tigeropen==1.1.5
tzlocal==1.5.1
urllib3==1.25.3
virtualenv==16.5.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.5
vnpy==2.0.4
websocket-client==0.56.0
yapf==0.26.0
zipp==0.5.1
